Refactor autoalloc tests
